+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/controller/GeneralSettingsController.java
@@ -136,7 +136,6 @@ public class GeneralSettingsController {
         command.setShowRememberMe(settingsService.isShowRememberMe());
         command.setPublishPodcast(settingsService.isPublishPodcast());
         command.setSearchMethodLegacy(settingsService.isSearchMethodLegacy());
-        command.setAnonymousTranscoding(settingsService.isAnonymousTranscoding());
 
         // Extensions and shortcuts
         command.setMusicFileTypes(settingsService.getMusicFileTypes());
@@ -154,7 +153,6 @@ public class GeneralSettingsController {
 
         // for view page control
         command.setUseRadio(settingsService.isUseRadio());
-        command.setUseSonos(settingsService.isUseSonos());
         User user = securityService.getCurrentUser(request);
         command.setShowOutlineHelp(outlineHelpSelector.isShowOutlineHelp(request, user.getUsername()));
         toast.ifPresent(command::setShowToast);
@@ -225,8 +223,6 @@ public class GeneralSettingsController {
         settingsService
                 .setSearchMethodChanged(settingsService.isSearchMethodLegacy() != command.isSearchMethodLegacy());
 
-        settingsService.setAnonymousTranscoding(command.isAnonymousTranscoding());
-
         // Extensions and shortcuts
         settingsService.setMusicFileTypes(command.getMusicFileTypes());
         settingsService.setVideoFileTypes(command.getVideoFileTypes());
@@ -245,7 +241,6 @@ public class GeneralSettingsController {
 
         // for view page control
         settingsService.setUseRadio(command.isUseRadio());
-        settingsService.setUseSonos(command.isUseSonos());
         if (!isReload) {
             redirectAttributes.addFlashAttribute(Attributes.Redirect.TOAST_FLAG.value(), true);
         }