+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/controller/UserSettingsController.java
@@ -25,6 +25,7 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -116,13 +117,13 @@ public class UserSettingsController {
         }
 
         command.setLdapEnabled(settingsService.isLdapEnabled());
-        command.setUsers(securityService.getAllUsers());
+        command.setUsers(securityService.getAllUsers().stream()
+                .filter(u -> !User.USERNAME_GUEST.equals(u.getUsername())).collect(Collectors.toList()));
         command.setAllMusicFolders(musicFolderService.getAllMusicFolders());
         command.setTranscodingSupported(transcodingService.isTranscodingSupported(null));
 
         // for view page control
         command.setUseRadio(settingsService.isUseRadio());
-        command.setUseSonos(settingsService.isUseSonos());
         toast.ifPresent(command::setShowToast);
         command.setShareCount(shareService.getAllShares().size());