+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/service/upnp/processor/UpnpProcessorUtil.java
@@ -24,14 +24,14 @@ import java.util.List;
 import java.util.ResourceBundle;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-import com.tesshu.jpsonic.dao.MusicFolderDao;
 import com.tesshu.jpsonic.domain.JpsonicComparators;
 import com.tesshu.jpsonic.domain.MediaFile;
 import com.tesshu.jpsonic.domain.MusicFolder;
 import com.tesshu.jpsonic.domain.Player;
-import com.tesshu.jpsonic.domain.User;
 import com.tesshu.jpsonic.service.JWTSecurityService;
 import com.tesshu.jpsonic.service.MusicFolderService;
+import com.tesshu.jpsonic.service.PlayerService;
+import com.tesshu.jpsonic.service.SecurityService;
 import com.tesshu.jpsonic.service.SettingsService;
 import com.tesshu.jpsonic.service.TranscodingService;
 import com.tesshu.jpsonic.util.StringUtil;
@@ -51,29 +51,33 @@ public class UpnpProcessorUtil {
 
     private final SettingsService settingsService;
     private final MusicFolderService musicFolderService;
-    private final JWTSecurityService securityService;
+    private final SecurityService securityService;
+    private final JWTSecurityService jwtSecurityService;
+    private final PlayerService playerService;
     private final TranscodingService transcodingService;
-    private final MusicFolderDao musicFolderDao;
     private final JpsonicComparators comparators;
 
-    public UpnpProcessorUtil(SettingsService ss, MusicFolderService mfs, JpsonicComparators c, JWTSecurityService jwt,
-            TranscodingService ts, MusicFolderDao md) {
+    public UpnpProcessorUtil(SettingsService ss, MusicFolderService mfs, SecurityService securityService,
+            JpsonicComparators c, JWTSecurityService jwt, PlayerService playerService, TranscodingService ts) {
         settingsService = ss;
         musicFolderService = mfs;
-        securityService = jwt;
+        this.securityService = securityService;
+        jwtSecurityService = jwt;
         comparators = c;
-        musicFolderDao = md;
+        this.playerService = playerService;
         transcodingService = ts;
     }
 
     public UriComponentsBuilder addJWTToken(UriComponentsBuilder builder) {
-        return securityService.addJWTToken(builder);
+        return jwtSecurityService.addJWTToken(builder);
     }
 
     public String createURIStringWithToken(UriComponentsBuilder builder, MediaFile song) {
         String token = addJWTToken(builder).toUriString();
         if (settingsService.isUriWithFileExtensions() && !StringUtils.isEmpty(song.getFormat())) {
-            token = token.concat(".").concat(song.getFormat());
+            Player player = playerService.getGuestPlayer(null);
+            String fmt = transcodingService.getSuffix(player, song, null);
+            token = token.concat(".").concat(fmt);
         }
         return token;
     }
@@ -82,11 +86,8 @@ public class UpnpProcessorUtil {
         return addJWTToken(builder).build().encode().toUri();
     }
 
-    public List<MusicFolder> getAllMusicFolders() {
-        if (settingsService.isDlnaGuestPublish()) {
-            return musicFolderDao.getMusicFoldersForUser(User.USERNAME_GUEST);
-        }
-        return musicFolderService.getAllMusicFolders();
+    public List<MusicFolder> getGuestMusicFolders() {
+        return musicFolderService.getMusicFoldersForUser(securityService.getGuestUser().getUsername());
     }
 
     public String getBaseUrl() {
@@ -115,8 +116,9 @@ public class UpnpProcessorUtil {
         return resourceBundle.getString(key);
     }
 
-    public boolean isDlnaGenreCountVisible() {
-        return settingsService.isDlnaGenreCountVisible();
+    public boolean isGenreCountAvailable() {
+        return settingsService.isDlnaGenreCountVisible()
+                && getGuestMusicFolders().equals(musicFolderService.getAllMusicFolders());
     }
 
     public boolean isProhibitSortVarious() {